% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{onload_calls_httr_POST_linter}
\alias{onload_calls_httr_POST_linter}
\title{Flag httr::POST calls inside .onLoad hooks
...}
\usage{
onload_calls_httr_POST_linter()
}
\description{
Flag httr::POST calls inside .onLoad hooks
...
}
