% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{lintrsecurity_linters}
\alias{lintrsecurity_linters}
\title{All lintrsecurity linters}
\usage{
lintrsecurity_linters()
}
\value{
A named list of linter functions.
}
\description{
Returns all stable lintrsecurity linters as a named list suitable for
passing to \code{lintr::lint()} or \code{lintr::lint_package()}.
}
