% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{onload_calls_httr_POST_linter}
\alias{onload_calls_httr_POST_linter}
\title{Flag httr::POST() call inside .onLoad or .onAttach hook}
\usage{
onload_calls_httr_POST_linter()
}
\value{
A linter function of class \code{linter}.
}
\description{
\code{httr::POST()} calls inside \code{.onLoad} or \code{.onAttach} make HTTP POST requests,
sending data to an external system whenever a package is loaded (e.g., via
\code{library()}) or attached to the R search path.
}
\details{
HTTP POST requests may be used to send data to external systems. For example,
they can send the contents of the \verb{~/.ssh} directory to an attacker. See
MITRE ATT&CK
\href{https://attack.mitre.org/techniques/T1552/001/}{MITRE ATT&CK T1552 (Unsecured Credentials: Credentials In Files)}.
}
\examples{
\dontrun{
lintr::lint(
  text = '.onLoad <- function(libname, pkgname) { httr::POST() }',
  linters = onload_calls_httr_POST_linter()
)
}

}
