% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload_calls_system_linter.R
\name{onload_calls_system_linter}
\alias{onload_calls_system_linter}
\title{Flag system() or system2() calls inside .onLoad or .onAttach.}
\usage{
onload_calls_system_linter()
}
\value{
A linter function of class \code{linter}.
}
\description{
Flag system() or system2() calls inside .onLoad or .onAttach.
}
\details{
\strong{Lint message:} system() or system2() inside .onLoad or .onAttach hooks execute
arbitrary shell commands automatically when a package is loaded.

\strong{Severity:} \code{warning}

\strong{MITRE ATT&CK:} \href{https://attack.mitre.org/techniques/T1059/004/}{T1059.004}, \href{https://attack.mitre.org/techniques/T1195/002/}{T1195.002}

For full documentation of the targeted vulnerability, see the \href{https://tylerjssmith.github.io/pkgaudit}{pkgaudit documentation}.
}
\examples{
\dontrun{
lintr::lint(
  text    = ".onLoad <- function(l, p) { system(\"id\") }",
  linters = onload_calls_system_linter()
)
}

}
