% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{onload_calls_system_linter}
\alias{onload_calls_system_linter}
\title{Flag system() calls inside .onLoad hooks
...}
\usage{
onload_calls_system_linter()
}
\description{
Flag system() calls inside .onLoad hooks
...
}
